pyproject.toml
[tool.poetry]
name = "custom_chatbot_server"
version = "0.1.0"
description = ""
authors = ["Masahiro Nishimi <mah@edged-edge.jp>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.21.1"
fastapi = "^0.95.0"
langchain = "^0.0.129"
openai = "^0.27.2"
gunicorn = "^20.1.0"
pydantic = "^1.10.7"
starlette = "^0.26.1"
faiss-cpu = "^1.7.4"
python-dotenv = "^1.0.0"
pypdf = "^3.9.1"
pycryptodome = "^3.18.0"
ipdb = "^0.13.13"
tiktoken = "^0.4.0"
ipykernel = "^6.23.1"
transformers = "^4.30.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
